fix: auth type error 수정
